v8.1 - errei fui mlk

+++ sql/data.sql
@@ -78,8 +78,8 @@ CREATE TABLE IF NOT EXISTS empresas (
     nome_emprsarial VARCHAR(255),
     natureza_juridica VARCHAR(15),
     qual_responsavel INT,
-    capital_social NUMERIC(15,2),
-    porte DOUBLE PRECISION,
+    capital_social DOUBLE PRECISION,
+    porte INT,
     ente_federativo VARCHAR(255),
     FOREIGN KEY (natureza_juridica) REFERENCES natureza(codigo)
 );
